requirements.txt
albumentations==1.0.0
einops==0.8.1
h5py==3.8.0
hdf5plugin==4.1.1
joblib==1.0.1
matplotlib==3.3.4
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.5.0
PyYAML==5.3.1
scipy==1.6.1
timm==0.4.12
torch==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.56.0
thop
fvcore
tensorboard
tensorboardX
torch-scatter